requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
exceptiongroup==1.1.3
fastapi==0.103.2
frozenlist==1.4.0
h11==0.14.0
idna==3.4
multidict==6.0.4
packaging==23.2
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
yarl==1.9.2
